Prevent extra calls to server
Hi,

We had some issues with the picker. For some reason we got a lot of 404 errors. It turned out the server gets a lot of entity requests with no id in there. 

So I took a look at the js en changed it a little.
If the id collection has no id's there will be a request to the server. First check if there are any id's, then make a request to the server.